nargo_cli

Module cli

Modules

check_cmd 🔒

compile_cmd

dap_cmd 🔒

debug_cmd 🔒

doc_cmd 🔒

execute_cmd 🔒

expand_cmd 🔒

export_cmd 🔒

fmt_cmd 🔒

fuzz_cmd 🔒

generate_completion_script_cmd 🔒

info_cmd 🔒

init_cmd 🔒

interpret_cmd 🔒

Use the SSA Interpreter to execute a SSA after a certain pass.

lsp_cmd 🔒

new_cmd 🔒

test_cmd 🔒

Structs

NargoCli 🔒

NargoConfig

PackageOptions 🔒

Options for commands that work on either workspace or package scope.

Enums

LockType 🔒

What kind of lock to take out on the (selected) workspace members.

NargoCommand 🔒

Constants

GIT_HASH 🔒

IS_DIRTY 🔒

NARGO_VERSION 🔒

Statics

VERSION_STRING 🔒

Traits

WorkspaceCommand 🔒

Commands that can execute on the workspace level, or be limited to a selected package.

Functions

compile_with_maybe_dummy_workspace 🔒

“with_workspace”, but use a dummy workspace when ‘debug_compile_stdin’ is enabled

lock_workspace 🔒

Lock the (selected) packages in the workspace. The lock taken can be shared for commands that only read the artifacts, or exclusive for the ones that (might) write artifacts as well.

read_workspace 🔒

Read a given program directory into a workspace.

start_cli 🔒

with_workspace 🔒

Find the root directory, parse the workspace, lock the packages, then execute the command.