nargo_cli::cli

Module check_cmd

Structs

CheckCommand 🔒

Check a local package and all of its dependencies for errors

Functions

check_package 🔒

Evaluates the necessity to create or update Prover.toml and Verifier.toml based on the allow_overwrite flag and files’ existence. Returns true if any file was generated or updated, false otherwise.

create_input_toml_template 🔒

Generates the contents of a toml file with fields for each of the passed parameters.

run 🔒