nargo_cli::cli::check_cmd
Struct CheckCommand

pub(crate) struct CheckCommand {
    pub(super) package_options: PackageOptions,
    pub(super) allow_overwrite: bool,
    compile_options: CompileOptions,
    show_program_hash: bool,
}

Check a local package and all of its dependencies for errors

Fields
package_options: PackageOptionsallow_overwrite: bool

Force overwrite of existing files

compile_options: CompileOptionsshow_program_hash: bool

Just show the hash of each packages, without actually performing the check.