nargo_cli::cli

Module doc_cmd

Structs

DocCommand 🔒

Builds documentation for the specified package or workspace.

Functions

collect_dependencies 🔒

package_crate 🔒

Returns the Crate item for the given package, together with all of its dependencies in a HashMap that maps a dependency’s root file to its crate.

run 🔒