nargo_cli::cli::interpret_cmd
Function compile_into_program

fn compile_into_program(
    file_manager: &FileManager,
    parsed_files: &ParsedFiles,
    workspace: &Workspace,
    package: &Package,
    options: &CompileOptions,
) -> CompilationResult<(Program, Abi)>

Compile the source code into the monomorphized AST, which is one step before SSA passes.

This isn’t exposed through the nargo library operations at the moment, so this is a bit of copy pasting from the functions that normally produce an artifact.