nargo_cli::cli::test_cmd::formatters
Trait Formatter

pub(crate) trait Formatter:
    Send
    + Sync
    + RefUnwindSafe {
    // Required methods
    fn package_start_async(
        &self,
        package_name: &str,
        test_count: usize,
    ) -> Result<()>;
    fn package_start_sync(
        &self,
        package_name: &str,
        test_count: usize,
    ) -> Result<()>;
    fn test_start_async(&self, name: &str, package_name: &str) -> Result<()>;
    fn test_end_async(
        &self,
        test_result: &TestResult,
        file_manager: &FileManager,
        show_output: bool,
        deny_warnings: bool,
        silence_warnings: bool,
    ) -> Result<()>;
    fn test_end_sync(
        &self,
        test_result: &TestResult,
        current_test_count: usize,
        total_test_count: usize,
        file_manager: &FileManager,
        show_output: bool,
        deny_warnings: bool,
        silence_warnings: bool,
    ) -> Result<()>;
    fn package_end(
        &self,
        package_name: &str,
        test_results: &[TestResult],
        file_manager: &FileManager,
        show_output: bool,
        deny_warnings: bool,
        silence_warnings: bool,
    ) -> Result<()>;
}

A formatter for showing test results.

The order of events is:

Compilation of all packages happen (in parallel). There’s no formatter method for this.

If compilation is successful, one package_start_async for each package.

For each test, one test_start_async event (there’s no test_start_sync event because it would happen right before test_end_sync)

For each package, sequentially:

A package_start_sync event

One test_end event for each test

A package_end event

The reason we have some sync and async events is that formatters that show output to humans rely on the sync events to show a more predictable output (package by package), and formatters that output to a machine-readable format (like JSON) rely on the async events to show things as soon as they happen, regardless of a package ordering.

Required Methods

fn package_start_async( &self, package_name: &str, test_count: usize, ) -> Result<()>

fn package_start_sync( &self, package_name: &str, test_count: usize, ) -> Result<()>

fn test_start_async(&self, name: &str, package_name: &str) -> Result<()>

fn test_end_async( &self, test_result: &TestResult, file_manager: &FileManager, show_output: bool, deny_warnings: bool, silence_warnings: bool, ) -> Result<()>

fn test_end_sync( &self, test_result: &TestResult, current_test_count: usize, total_test_count: usize, file_manager: &FileManager, show_output: bool, deny_warnings: bool, silence_warnings: bool, ) -> Result<()>

fn package_end( &self, package_name: &str, test_results: &[TestResult], file_manager: &FileManager, show_output: bool, deny_warnings: bool, silence_warnings: bool, ) -> Result<()>