nargo_cli::cli
Trait WorkspaceCommand

trait WorkspaceCommand {
    // Required methods
    fn package_selection(&self) -> PackageSelection;
    fn lock_type(&self) -> LockType;
}

Commands that can execute on the workspace level, or be limited to a selected package.

Required Methods

fn package_selection(&self) -> PackageSelection

Indicate which package the command will be applied to.

fn lock_type(&self) -> LockType

The kind of lock the command needs to take out on the selected packages.