nargo_cli::cli

Module compile_cmd

Structs

CompileCommand

Compile the program and its secret execution trace into ACIR format

Functions

compile_contracts 🔒

Compile the given contracts in the workspace.

compile_programs 🔒

Compile the given binary packages in the workspace.

compile_workspace 🔒

Compile binary and contract packages. Returns the merged warnings or errors.

compile_workspace_full

Parse and compile the entire workspace, then report errors. This is the main entry point used by all other commands that need compilation.

parse_workspace

Parse all files in the workspace.

run 🔒

save_contract 🔒

watch_workspace 🔒

Continuously recompile the workspace on any Noir file change event.